Use current, but pinned versions of pip, setuptools and wheel
This patch provides a convenience script to check for the current
version of any package on PyPI, then output it in various ways.

This script is used in the SHA updating script in order to provide
a current set of critical packages to ensure that each SHA bump
includes an update to a current version of pip, setuptools and
wheel but also to ensure that they are pinned to a specific version
with this particular set of packages.

This ensures that we keep current with these packages as they
change, but also ensures that the versions tested for each tag are
the versions used forever.

The patch also ensures that any package installed by pip is upgraded
to the expected versions.

This is a combined backport of:
 - https://review.openstack.org/284701
 - https://review.openstack.org/284977
 - https://review.openstack.org/287700
 - https://review.openstack.org/287739

Backport notes:
 - https://review.openstack.org/283039 is reverted as the pinning
   is now handled properly elsewhere.
 - The pip pinning done via global-requirements in
   https://review.openstack.org/270312 is removed.
 - The versions pinned are the versions used at the time of the
   last rpc-repo build. This is to ensure that this tag implements
   the same version of all the packages used as the time the tag
   repo was built.
 - The SHA for the requirements repository is updated in order to
   include the caps for osprofiler and testtools.

Closes-Bug: 1545505
Change-Id: I61795b3afb4804060d494a08975c10adcf52f468
(cherry picked from commit 6ea8e986f1bc826ffd54477eddc1ea6f04b8ea53)