Add haproxy-install to setup-infrastructure.yml
Right now users who are using haproxy rather than a physical
load balancer need to run the haproxy-install.yml playbook,
but it is not included in setup-everything. It is not very clearly
documented that running this seperate playbook is necessary, since
the haproxy configuration is listed as (optional) in the installation
guide. Also, if the user has aphysical load balancer they will not
configure haproxy_hosts in /etc/openstack_deploy, so this playbook
will be skipped anyways.

Change-Id: I368adf5a19ba865e4e21bc8bcde096018bed8389
Partial-bug: #1552362
(cherry picked from commit f5333abb988867f271224bbf938f3cf201f711f0)