Clarify HAProxy configuration necessity
This was a little roadblock that I ran into when deploying
using openstack-ansible recently. The documentation doesn't
indicate that configuration of a load balancer is explicitly
required, and since the HAProxy installation playbook didn't
run automatically on setup-everything.yml the errors that I
got didn't lead me to figuring out that HAProxy was the
missing piece.

Change-Id: I8c15910f19e2be8ece3bac98bcbf82eb49045bf8
Partial-bug: #1552362