Test basic structure of dynamic inventory output
This is intended as a building block for future refactoring of the dynamic
inventory script.

Tests validate the output:
 * meets basic content requirements imposed by Ansible
 * contains all expected OSA host groups
 * contains only the OSA expected host groups

Additionally, the iteration count for the duplicate IP test has been changed
from 25 to 100 since the check execution time is not adversely effected by the
difference and the original author had wanted 100 iterations.

Change-Id: I17971a35b69ebe4aa4dcdbc60210a5bbbc6764a3