Ensure that the repo-server uses upper-constraints
In order to ensure that builds are repeatable (ie the result is
always the same when building from a tag), the repo server's
packages must make use of OpenStack's upper-constraints file
for its own packages.

This patch ensures that this happens, and implements the
additional changes required for it to succeed.

This is a combined backport of:
 - https://review.openstack.org/284894
 - https://review.openstack.org/286923
 - https://review.openstack.org/286625 (partial)

Closes-Bug: 1545505
Change-Id: I157fd3f4e116afa886a27fc1f1422a3a04321c0a