Bypass repo server if missing pip upgrade packages
The function get_pip upgrades the pip packages for pip. If OSA is
updated to require newer versions these may not be present on the repo
server and this will cause bootstrap-ansible.sh to fail.

Update the get_pip function to fallback on calling pip with the
--isolated flag when pip is installed and the pip packages are being
upgraded.

Change-Id: Ifa644b9bf6aa8b20028eb0257c40fe20f3ab1a3b
(cherry picked from commit 721ef48fbf66e375af583710117f5c3bfbace9b9)