Check correct minimum size for data disk
Currently we choose the largest available disk as the data disk, however
this may be too small to be useful. This patch checks disqualifies disks
that are smaller than the minimum.

It also sets the minimum to the proper value of 60GB free on the root
disk or 60GB available on the designated secondary disk. The previous
check was 60GB on root and 80GB on the secondary disk which is not
the intented check as noted in the documentation.

Change-Id: Id225dc55068e1eed2e19e1e27b0f626163f03db2