Cap bashate and fix linters tests
Bashate is pinned to ensure that for the rest of the cycle, and
once this becomes a stable branch, there are no code changes
needed for updates to the linting tool.

Also, the tox configuration is updated to ensure that it consumes
the dev-requirements file for all lint testing.

Change-Id: I78056d1d22228f073c44e9ee4e78534a4010effe
(cherry picked from commit 13bea329896c82a8352633038090f31e09ef7afd)