Set apt sources to use a configured list of components
re: http://lists.openstack.org/pipermail/openstack-dev/2016-February/086000.html

Ubuntu has 4 different 'components' - main, universe, multiverse and
restricted:
 - Main: Officially supported software.
 - Restricted: Supported software that is not available under a completely
               free license.
 - Universe: Community maintained software, i.e. not officially supported
             software.
 - Multiverse: Software that is not free.

Practically speaking there should be nothing particularly useful to
OpenStack-Ansible in Restricted or Multiverse - it's mostly software for
desktop users.

This patch introduces a new variable 'lxc_container_template_apt_components'
which is a list of the components to configure in the apt sources list. The
default list includes all the components to ensure that existing deployments
do not have any changes to the configuration implemented.

This patch also introduces a new variable 'bootstrap_host_apt_components'
which is a list of the components to configure in the apt sources list for
the AIO and gate testing specifically. The default list does not include the
unnecessary components.

Backport notes:

This is a combined backport of https://review.openstack.org/278033 and
https://review.openstack.org/278037, but with the lxc_hosts role defaults
adjusted in order to protect existing production deployments.

Change-Id: I0eef6454d273f5ba1977a68151fbf6103ff3ed8d