Make the purpose of the example user_variables file clearer
This patch changes the example user_variables.yml file to ensure
that it's clear that the contents are for convenience and that
none of the settings are required to be set.

Where some settings were implemented in the file but were not
properly defaulted in the group_vars, this has been corrected.

Change-Id: Ib5e901ebca30fa1810d7801caafda8287c911545