Gate: Use OpenStack-CI apt mirrors
re: http://lists.openstack.org/pipermail/openstack-dev/2016-February/086295.html

OpenStack-CI has now implemented their own APT mirrors which are updated and
synchronised properly to prevent issues when the update takes place between
the time an instance does an 'apt-get update' and an 'apt-get install'.

Now that this infrastructure is in place, we can remove our logic to determine
the optimal apt source for gating and simply fall back to the default AIO
behaviour which is to use whatever is configured on the host.

In order to do this successfully the AIO needs to ensure that it has the
appropriate configurations in place to ignore unauthenticated packages.

This patch implements a full propogation of all files in
/etc/apt/apt.conf.d/ from the host to all containers in order to ensure
that any configurations implemented by OpenStack-Infra are used from
now on in our Gate jobs.

Change-Id: Id6c2be3cf57b7c49744156a943f9653c1575e3dd