Move inventory environment loading to function
Originally the inventory environment was loaded as part of the `main`
function. While this worked, it was difficult to isolate environment
information when using it for testing.

The loading process for the environment is now in a standalone function,
with basic tests around it.

In order to facilitate testing the `load_environment` function directly,
the Python path is altered in the test script to import the
dynamic_inventory.py script. This change is also useful for future tests
targeted at specific functions.

Change-Id: Ic24a8bc068f39cbd56ec8563d7f0bb40690f63a3