Fail when ansible-galaxy returns error
Remove the --ignore-errors flag from the ansible-galaxy command used to
pull in the roles required by the project. If a role fails to be
installed, ignoring the error only moves the failure from
bootstrap-ansible.sh to the playbooks.

Adjust the ordering of the roles in the requirements file so that
dependencies are installed before they are required. This is necessary
because the roles are not on Ansible Galaxy.

Modify tox.ini so that the functional testing does not ignore errors.

Change-Id: Id68affada5cf2380b0388d45a9baf6b4839824c7