Adjust swift plays to use unified os-swift role
In https://review.openstack.org/#/c/293536 os-swift-sync was merged into
os-swift. The playbooks in openstack-ansible need to be adjusted to set
the specific vars so that only the required tasks are run.

In order to speed up the gate run, we adjust the os-swift-install task
to run the os-swift role once, but have the same end result as before.

Change-Id: I3e3891333229be0fb933479837a2fca451eb7474
Depends-On: If01d8a3aa29fc7637aecb92e61929523af08103d