Add installation support for os_ironic
This patch adds the initial support for the ironic role in
openstack-ansible, but leaves ironic unconfigured and not
installed by default.

This is a combined backport of:
 - 35ed804e3a39db3ec0fe801e0de4f85358594e63 (initial integration work)
 - 0d9530c9fd64a590ca81c1fdd87c26731c6ada44 (secrets addition)
 - aa1f09f12a6902e90d9b78e4aca6b2621f9ec222 (documentation)

Backport note:
The instrumentation of Ironic to be included in the AIO deployment
has specifically been disabled in the backport. Ironic is an
experimental feature at this time and has no functional tests which
are usable in an AIO.

For the sake of being able to use it for experimental test
environments all the instrumentation is available to deploy it
and it is included in the setup-openstack play list.

Change-Id: Id9f01deb5c46ee2186b9c41c7f88205560b5f437
Depends-On: Ide66c7ee59192ac441ac2919028eca0ad665ceea
Depends-On: I590f5ade90b3e37af7f1b8ee333000d4f993f8c5
Partially-implements: blueprint role-ironic
Signed-off-by: Michael Davies <michael@the-davies.net>
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>
(cherry picked from commit 35ed804e3a39db3ec0fe801e0de4f85358594e63)