Disable Neutron port security driver on upgrades
When upgrading from Kilo to Liberty, networks existing prior to the
upgrade prevented new VMs from booting. This was due to the introduction
of the port_security extension to Neutron's defaults in the
openstack-ansible Liberty release.

This change disables the port_security driver in environments upgrading
from Kilo, with documentation informing deployers of the change.

Greenfield Liberty environments do not experience this behavior, as the
port_security driver is enabled for all networks created there.

Upstream Neutron do not currently provide a mechanism for adding
port_security binding information to pre-existing networks, nor does it
fail gracefully. Hence, this patch is deemed the least invasive way to
work around the issue for openstack-ansible.

Change-Id: Ib0d9eeae3766626186586c22875a2c1acf259e03