Remove the repo clone mirror play
This play is no longer used and is a hold over from the time
when the OpenStack-Ansible project came from the Rackspace
Private Cloud product. Being this is not a play used by
the core project it should be removed.

This change also resolves an issue/misconseption that using
the clone play is a better/more stable way to build/deploy
an OpenStack cloud using OpenStack-Ansible.

Change-Id: Icc4d3d01ecd0e3eb2de1ad54b9bff1ca1662beb4
Closes-Bug: #1559619
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>