Add group_vars for swift_remote_all hosts
When expanding hostvars for a different host, ansible will attempt to
expand the vars that are passed through a filter (netloc or netorigin
for example). This causes the ansible_remote_all hosts to fail out
during the SSH key sync phase.

In order to get this to work successfully we need to define the vars
that are being expanded (even though they aren't technically required by
the remote_hosts or swift plays) so that the remote hosts can access the
local hosts' hostvars.

Change-Id: I7f74a8404a8c805cdaf78596b1b32e7de95238c0