If a global pip.conf file exists, let the AIO use it for containers
OpenStack-CI implements a global pip configuration which configures
the use of the OpenStack pypi and wheel mirrors.

This patch implements a check for a global pip configuration file,
then adds this file to the list of files to be copied into the
container cache so that all containers have the configuration.

While most containers end up with a localised pip configuration
which locks them down to the repo server (which will ignore the
glocal config), this is very useful for the repo server which
does not get locked down.

Change-Id: I1058b68f2281c5152fcd4b880fa21121716bc05c