Ensure the OpenStack gate has access to the logs
When the gate fails, for any reason, its possible the logs generated
by the services are not accessible by the jenkins worker which does
cause problems when the logs are uploaded for analysis. To ensure the
gate always has access to the logs a trap was added to the gate script
which is used to guarantee the gate has access to the logs by recursively
setting the mode to 0777 on all files within the "/openstack/logs"
directory.

The function "gate_job_exit_tasks" has been added to the
script-libraries.sh file

Change-Id: I3eece015060dd99aceb7a39ca62c5cb1a0162421
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>
(cherry picked from commit 6eb3c34a17cf2ad0ceae8cdff55569b999369f13)