Set lxc_container_caches not to use repo_pip_default_index
The LXC container cache may happen to use the same web server
as that set for the repo_pip_default_index but this is not an
assumption that should be made so readily.

This patch purposefully decouples these.

Change-Id: I9f8901d8c7f48a1de0511eb0ba1795a1db28405c
(cherry picked from commit 6f9ef5f1686c8c4104ccbf789ddd7d88b6ec84d2)