Set default Glance store for the AIO back to Swift
In I02b69cf5985e8788513db58f7e8015f8135b9d58 the Glance Store
for Swift was appropriately configured to work with the Keystone
v3 API and in the new expected way for the Glance Swift Store to
be configured for Mitaka.

This patch returns the default AIO to use Swift as the backing
store for Glance.

Change-Id: Iec233f990bd032fb29f633d74137c29bc89f63f5
Closes-Bug: #1561947