Modify the haproxy play for ansible2 compat
The haproxy play fails when being executed under ansible 2.x
The failure is caused by the task include which reaches into
the hapeoxy role to rerun a template action when a conditional
is met. To make this play ansible 2.x compatible the include was
removed and the role was simply called a second time with a
provided set of variables which will accomplish the same thing.
Without this change the services.j2 template would require a
symlink in the playbook directory in order to allow ansible 2.x
to access the file.

Change-Id: I9cb93b2bcb7f1a4140a73e6e4cae38803b894a2c
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>