Provide Nova & Cinder the Glance API Servers
Currently the Nova and Cinder roles require the Glance API server
list, but use variables outside of their namespace.

This is not good practise. This sets up namespaced group_vars for
use by both roles.

The Cinder role already expects the var [1], but a follow-on patch
will be submitted to the Nova role to also adjust its pattern.

[1] https://review.openstack.org/290856

Change-Id: I31b06aad686fcc32d55c78147e3d5d099b8cc5b9