Restart rsyslogd after logrotate
The copytruncate directive in logrotate configuration truncates the log file so
that processes can keep using the same file descriptor to write to the log
file. The truncation, however, does not work well with rsyslog forwarding as
rsyslog fails to figure out the difference from last sent messages.

This change restarts rsyslogd in postrotate directive of logrotate such as
rsyslog will start with a new baseline and continue forwarding logs.

This change back ports the master change of https://review.openstack.org/300249

Change-Id: I37c2778a07a697cdeded82d5d7c269dd3cded85a
Closes-Bug: 1564654
Co-Authored-By: Shu Shen <sshen@siaras.com>