Ensure cleanup-rabbitmq-vhost is re-runnable
Currently the cleanup-rabbitmq-vhost playbook will always fail on a
second run since the rabbitmq_user_to_delete will not exist to add tags
to or to change its password.

Make use of the rabbitmq_user module and force the creation of this user
on each run so that its consistently available.

Change-Id: Ib4b907fcf2f7c62ba9e112bec95c0714332177c2