Fix failures in Juno to Kilo upgrade script
ba6255bc3c67beed6a2f3e8c4f0ad8758e51d16c introduces a bug which results
in the playbook lxc-containers-create.yml not being run on a number of
containers. The function run_lock in run-upgrade.sh prevents a playbook
from being run more than once. The playbook lxc-containers-create.yml is
run three times with --limit to control the order in which containers
are updated. The result is that only the first occurrance of
lxc-containers-create.yml is executed.

This commit creates three wrapper playbooks for
lxc-containers-create.yml so that run_lock treats them as separate
actions.

235c2b29be7fa4777a455377f401066e557c9f21 introduced a new playbook to
clean up MariaDB apt sources. The playbook targets the wrong directory,
'/etc/apt/repos.d/' when it should be '/etc/apt/sources.list.d'. In
addition, the playbook is run before all the new containers have been
created and so fails.

This commit moves the MariaDB clean up into juno-container-cleanup.sh so
that it is located with the other related clean up actions. As with the
other steps in that script, it now ignores failures when removing the
MariaDB source on the assumption that the failures are due to non-
existant containers. In addition the path to search is changed to
'/etc/apt/sources.list.d'. Finally the command is now also applied to
physical hosts in addition to containers.

Change-Id: Ia57d18e09a7df30bf9cbbb5d77177ae7e8003832