DOCS: Configuration section - cleanup
As per discussion in the OSA docs summit session, clean up
of installation guide. This fixes typos, minor RST mark up
changes, and passive voice.

This patch also merges a some of the sections into the larger
chapter. This is in an effort to remove multiple smaller
files.

This patch is the first of many to avoid major conflicts.

Combined backport of:
 - I6801653551e79eaf74f7a9add0030d7e3688194d
   (cherry picked from commit 43ff983e4d0100ed3b101377cd205767ca9d3655)
 - Iaf3d56751dadfaf02b66ac248e01fd425b28c777
   (cherry picked from commit 2c5edcf43ee0c03251f8234b5bad667197bd9fc2)
 - I5b6d540bfa691b36dbe9cf909de8e61affb0cd92
   (cherry picked from commit e78c9e3ac2fea33008fedbed15606707e7d033cc)

Co-Authored-By: Robb Romans <robb.romans@rackspace.com>
Change-Id: I5b6d540bfa691b36dbe9cf909de8e61affb0cd92