Fix container hostname for RFC 1034/1035
In lxc-container create this update will ensure we're using good
hostnames. This is required in some logging cases as well as SSL.
This will also allow a deployer to customize the domain name
based on their needs.

A new variable has been created called ``lxc_container_domain``
with a default value of "openstack.local". This change has no
upgrade impact as its being done as part of the post_tasks within
the container create play and any changes that need to happen
within an existing environment will be automatically taken care
of within impacting the functionality of an existing deployment.

Change-Id: Ib31a48dd480ecb376a6a8c5b35b09dfa5d2e58f6
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>