Refactor main inventory function for testability
This change alters the dynamic_inventory.py file's main function with
two main changes:

    * main now accepts arguments, which could be the ArgParse parsed
      arguments or a plain dictionary for testing
    * return output to then be printed, instead of just printing.

With these changes, testing inventory can now be done via Python imports
rather than a subprocess call. This make existing tests more clear, and
allows for using pdb when creating future patches. With the subprocess
method, a pdb in the dynamic_inventory.py would cause the test suite to
hang, since the subprocess was waiting for input.

Existing functionality is *not* changed - arguments and output are
simply being moved to make the code easier to test.

Change-Id: I3629c9d0704a7412832d7e898463d376b4e13317