Refactor run-playbooks
This changes the run playbooks script so that its using our core CLI
tools and is no longer a special case script running bits outside of
simply running the deployment tools as they've been designed.

The script is still executing all of the playbooks using the
"scripts-library" function ``install_bits`` so that we ensure gate
success by having a retry on a playbook execution in the case of transient
failure.

A notice has been added to the script whenever it exits notifying
the deployer that they should now work out of the playbooks directory
and never execute the run-playbooks.sh script ever again. Within the
notice links to our documentation have been added.

The ironic services were removed from the integrated gate because they're
not ready for general consumption and not being tested at this time because
no test configuration has been added into the tempest role or the
``run-tempest.sh`` script.

Change-Id: I064c1c480d261a885666b7dc3edc4cd0769876a6
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>