Add missing line number report, fix coverage dep
The coverage missing line number report will be helpful in the gate to
show exactly what lines were not exercised by the existing tests.

Also, when coverage was added to the test requirements file it appeared
to have broken the `inventory` tox environment because coverage was also
specified directly in the tox.ini file. Even if it does not break the
gate, being specified in both places is at best redundant.

Change-Id: Ic94313018f565e186a93c991906fbfc802a84034