Add docs for HAProxy ping checks
The HAProxy role configures a ping check that runs against a root
DNS server operated by RIPE. This may not be reachable in certain
deployments and some deployers might not want to ping this external
server repeatedly.

This patch adds documentation for deployers who want to change the
destination for ping checks or the count/frequency of these pings.

Closes-bug: 1573640

Change-Id: I9ab29b9df7ec35b197f203cb287402e860371c8e