Updates all SHAs for 13.0.2
This patch updates all the roles to the latest available stable
SHA's, updates all the OpenStack Service SHA's and also updates
the appropriate python requirements pins.

The wheel pin is permanently set to 0.29.0 in order to match the
upper-constraint in OpenStack Mitaka which remains static once a
branch becomes stable.

The sources-branch-updater script has been changed to ensure that
this constraint always remains constant.

Change-Id: I1be61fbbfb38764f95d8463a8c49d9a920bb8827