Updates all SHAs for 12.0.12
This patch includes updates of any changed paste, policy and rootwrap
configurations. It also includes updates to the pip, wheel and
setuptools pins.

The wheel pin is permanently set to 0.26.0 in order to match the
upper-constraint in OpenStack Liberty which remains static once a
branch becomes stable.

The sources-branch-updater script has been changed to ensure that
this constraint always remains constant.

The python package python-openstackclient is temporarily pinned to
1.7.2 in order to ensure that it can be co-installed with other
upper-constrained packages in the utility container. This pin can
be removed once https://review.openstack.org/312581 merges.

Change-Id: Ic3c688de7c547eda73bbd0b4893010e18f401a86