Updates all SHAs for 11.2.15
This patch includes updates of any changed paste, policy and rootwrap
configurations. It also includes updates to the pip, wheel and
setuptools pins.

The wheel pin is permanently set to 0.26.0 in order to ensure that
Kilo's version matches the upper-constraint in Liberty in the hope
that this improves the chances of upgrade success.

The sources-branch-updater script has been changed to ensure that
this constraint always remains constant.

Change-Id: I0ccbcb2d8e7ebbb692b1a042e4280f0f68c6cdcc