Limit galera container restarts during upgrade
There is a slight difference in the container interface template between
Kilo and Liberty. When the file is changed, a container restart is
notified. This can lead to all galera containers being restarted nearly
simultaneously with none of their mysql services coming back up in a
working state afterwards.

Exclude the first galera host from the setup-hosts playbook, and run the
lxc-container-create playbook specifically against it afterwards to
ensure that, in a deployment with multiple galera containers, there is
always a galera container that remains up during an upgrade.

Change-Id: I681f5b1de4645130f5e1a530655fdda6b16d8b9e