Set test python executable to python2
If a developer uses an operating system with Python 3 as a default, tox
will use the Python 3 interpreter to run the tests. However, this
introduces some errors since some standard library modules and classes
are changed or missing in Python 3. This will cause some false
failiures.

By enforcing the base python to Python 2, the tests are ensured to be
running on the same interpreter regardless of the operating system's
default.

Change-Id: Ied142fe44a38ee3119d02921ef680fd09ce62f0e