Restart services on pip package install
Service restart notifications are currently inconsistent in kilo.
Roles either rely on a forced reload of init scripts or watch for
changes in service configuration files, which may not happen.

This patch ensures that when there is a change to the project's pip
packages, its services are properly notified to restart.

Change-Id: I88c6fefbe51d803bfc4da79e7220f95978a134ec