Ignore failures when removing legacy apt sources
The directory '/etc/apt/sources.list.d' is not guaranteed to exist. When
the playbook mariadb-apt-cleanup.yml is run it will fail if the
directory is not there. This commit modifies the task to ignore failures
so that the upgrade script is able to complete regardless of whether or
not '/etc/apt/sources.list.d' exists or contains any files.

Change-Id: Ib87868897b70a5da127d6bfb9e35c8460a4281cd