Remove teardown.sh and update related docs
The teardown.sh script attempts to provide a convenience, but is often
left unmaintained. Because it is not maintained, it cannot fully tear
down an environment. In addition, maintaining it has been made much
harder since the introduction of independent role repositories - changes
within those individual repositories would need to either be torn down
in this script, or a tear down process in each role. Such a process
would again face issues with maintenance.

Also the teardown script has created bugs with installs started on a
'dirty' environment. These behaviors have been hard to diagnose and fix,
resulting in a lot of wasted time.

Because of these issues, this patch removes the teardown script
entirely, and alters the docs to recommend that AIOs are deployed to
virtual machines for easier tear down and redeploy, as well as advising
against any kind of production use.

Fixes-Bug: #1540531

Change-Id: Ida59bec0ff961424180940628b006d71e88e199b