Removed container_release property from environment files
This change removes the container_release property from environment
files used to generate inventory.  This change is because the
lxc_container_create role has been updated to build containers to the
same distribution and release as the host machine.

This change also removes the variable lxc_container_release from the
lxc-containers-create.yml play as it is no longer used by the
lxc_container_create role.

Change-Id: I7d2aea8f6548feb381bf2354965f05fa0a9a99b8
Depends-On: Iee304dd026e0865e0444259d2132122233d90f5f