Add release note for paramiko issue workaround
Change Ia7e0d64214cfd5412ec236ee4bdd6a076821d258 implemented the fix to resolve
the issue with the Paramkino 2.0 update, but did not include a release note to
be explicit that this is a known issue and to inform deployers how to work
around it.

This patch adds a release note to do that.

Change-Id: Iac3bca21fc336835ef1cad3a89d7cea0e4babaa8
Closes-Bug: #1577469