Make tox use python2.7 more specifically
When executing tox it tries to identify the appropriate Python 2
binary to use. When multiple Python 2.x binaries are available
then this may fail.

We specifically support the use of Python 2.7 so this patch
explicitly sets tox to make use of Python 2.7.

This resolves a situation where executing a tox test may result in
the error 'ERROR: InterpreterNotFound: python2'.

Change-Id: I647bc898343789e89963424033363a2d72b89c79